refactor: (b_route) Consolidate sensor integration into init and remove redundant sensor.py file
